Distance funtionality modified
